backend/requirements.txt
Django>=5.0.7
django-cors-headers>=4.4.0
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
